@@ -5,18 +5,20 @@
   "main": "index.js",
   "scripts": {
     "build": "tsc",
-    "start": "ts-node ./src/index.ts"
+    "start": "ts-node-esm ./src/index.ts"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "fastify": "^3.29.0",
-    "pg": "^8.7.3",
-    "ts-node": "^10.5.0",
-    "typescript": "^4.5.5"
+    "@fastify/cors": "^8.2.0",
+    "fastify": "^4.9.2",
+    "pg": "^8.8.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.3"
   },
   "devDependencies": {
     "@types/pg": "^8.6.5"
-  }
+  },
+  "type": "module"
 }
